KTOR-369 Verify invalid response parser error
